backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic==2.10.3
openai==1.58.1
tiktoken==0.8.0
websockets==14.1
httpx==0.28.1
PyYAML==6.0.2
openpyxl==3.1.5
faiss-cpu==1.9.0
numpy==1.26.4
mcp==1.12.4
